js is js my boi

+++ src/main/java/com/Wkrzyz/HouseHoldIncomeManager/model/UserGroup.java
@@ -14,7 +14,7 @@ public class UserGroup {
 
     private String name;
 
-    private Float balance;
+    private double balance;
 
     private String admin;
 
@@ -30,7 +30,7 @@ public class UserGroup {
     public UserGroup(){
 
     }
-    public UserGroup(Integer id, String name, Float balance, String admin, List<User> users, List<Transfer> userGroupTransfers) {
+    public UserGroup(Integer id, String name, double balance, String admin, List<User> users, List<Transfer> userGroupTransfers) {
         this.id = id;
         this.name = name;
         this.balance = balance;
@@ -55,11 +55,11 @@ public class UserGroup {
         this.name = name;
     }
 
-    public Float getBalance() {
+    public double getBalance() {
         return balance;
     }
 
-    public void setBalance(Float balance) {
+    public void setBalance(double balance) {
         balance = balance;
     }
